Temporarily disable Content-Type and Content-Encoding on HTTP sample
The samples have their own project files (.sln) which are built using
the published nuget packages. Since IoTHubMessage is being modified to
add a couple new API functions and it is still not released as a new
nuget package, the sample fails to compile (the new API functions are
not present on the currently published nugets).
These lines commented out on the HTTP sample shall be re-enabled after
the next release.