Remove hardcoded key name usage from iothub_messaging_ll.c
The service client for cloud-to-device messaging constructed a token authentication id using a hardcoded name of the policy owner.
Only policies with owner "iothubowner" would then work with this client.
This change removes this hard-coded name, consuming the policy owner name as provided by the user through the IoT Hub connection string.